signer/requirements.txt
psycopg2==2.9.9
loguru~=0.7.2
solana~=0.32.0
base58~=2.1.1
solders~=0.20.0
grpcio~=1.62.1
protobuf~=5.26.1
